+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Open/Close menu entry in smart imaging toolbar (right-click on IM icon)
+- SmartImaging > Set light source IDs (for click tool) plugin, allowing setting custom light source IDs instead of fluo1, 2, 3...
 
 ### Changed
 - Smart Imaging toolbar, IM icon now is meant to be right-clicked, and the play icon is the one providing the click-tool rescreening